[Bronze V] Title: 꼬마 정민, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/11382. 꼬마 정민/꼬마 정민.cc
@@ -4,7 +4,7 @@ using namespace std;
 int main(void){
     ios::sync_with_stdio(0);
 	cin.tie(0);
-    long long a,b,c;
+    long a,b,c;
     cin >> a >> b >> c;
     cout << a+b+c;
 }
